script: Don't delete system images when starting flynn-host

+++ host/cli/destroy-volumes.go
@@ -18,8 +18,9 @@ func init() {
 usage: flynn-host destroy-volumes [options]
 
 options:
-  --volpath=PATH     directory to create volumes in [default: /var/lib/flynn/volumes]
-  --include-data     actually destroy data in backends *this is dangerous* [default: false]
+  --volpath=PATH         directory to create volumes in [default: /var/lib/flynn/volumes]
+  --include-data         actually destroy data in backends *this is dangerous* [default: false]
+  --keep-system-images   don't destroy system images
 
 Destroys all data volumes on this host.  This is a dangerous operation: data
 may be permanently discarded.
@@ -33,6 +34,9 @@ it will simply be removed.  In this case, data will remain behind in the backend
 storage engines (and eventually require manual cleanup), but the flynn-host
 daemon's next launch will still be like a fresh launch.
 
+If '--keep-system-images' is set, system image volumes are not destroyed and
+the volume database is not removed.
+
 Major features of backends will not be adjusted (i.e., zfs datasets will be
 destroyed, but zpools will not be touched).`)
 }
@@ -45,6 +49,7 @@ func runVolumeDestroy(args *docopt.Args) error {
 
 	volPath := args.String["--volpath"]
 	includeData := args.Bool["--include-data"]
+	keepSystemImages := args.Bool["--keep-system-images"]
 
 	volumeDBPath := filepath.Join(volPath, "volumes.bolt")
 
@@ -64,18 +69,20 @@ func runVolumeDestroy(args *docopt.Args) error {
 	} else if includeData == false {
 		fmt.Println("'--include-data' not specified; leaving backend data storage intact.")
 	} else {
-		if err := destroyVolumes(vman); err != nil {
+		if err := destroyVolumes(vman, keepSystemImages); err != nil {
 			fmt.Printf("%s\n", err)
 			allVolumesDestroyed = false
 		}
 	}
 
-	// remove db file
-	if err := os.Remove(volumeDBPath); err != nil {
-		fmt.Printf("could not remove volume state db file %q: %s.\n", volumeDBPath, err)
-		shutdown.ExitWithCode(5)
+	if !keepSystemImages {
+		// remove db file
+		if err := os.Remove(volumeDBPath); err != nil {
+			fmt.Printf("could not remove volume state db file %q: %s.\n", volumeDBPath, err)
+			shutdown.ExitWithCode(5)
+		}
+		fmt.Printf("state db file %q removed.\n", volumeDBPath)
 	}
-	fmt.Printf("state db file %q removed.\n", volumeDBPath)
 
 	// exit code depends on if all volumes were destroyed successfully or not
 	if includeData && !allVolumesDestroyed {
@@ -106,24 +113,30 @@ func loadVolumeState(volumeDBPath string) (*volumemanager.Manager, error) {
 	return vman, nil
 }
 
-func destroyVolumes(vman *volumemanager.Manager) error {
+func destroyVolumes(vman *volumemanager.Manager, keepSystemImages bool) error {
 	someVolumesNotDestroyed := false
 	var secondPass []string
-	for volID := range vman.Volumes() {
-		fmt.Printf("removing volume id=%q... ", volID)
-		if err := vman.DestroyVolume(volID); err == nil {
+	for id, vol := range vman.Volumes() {
+		if keepSystemImages && vol.Info().Meta["flynn.system-image"] == "true" {
+			continue
+		}
+		fmt.Printf("removing volume id=%q... ", id)
+		if err := vman.DestroyVolume(id); err == nil {
 			fmt.Println("success")
 		} else if zfs.IsDatasetHasChildrenError(err) {
 			fmt.Println("has children, coming back to it later")
-			secondPass = append(secondPass, volID)
+			secondPass = append(secondPass, id)
 		} else {
 			fmt.Printf("error: %s\n", err)
 			someVolumesNotDestroyed = true
 		}
 	}
-	for volID := range vman.Volumes() {
-		fmt.Printf("removing volume id=%q... ", volID)
-		if err := vman.DestroyVolume(volID); err == nil {
+	for id, vol := range vman.Volumes() {
+		if keepSystemImages && vol.Info().Meta["flynn.system-image"] == "true" {
+			continue
+		}
+		fmt.Printf("removing volume id=%q... ", id)
+		if err := vman.DestroyVolume(id); err == nil {
 			fmt.Println("success")
 		} else {
 			fmt.Printf("error: %s\n", err)
